+++ docs/changelog.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Completed and revised docstrings in `stepup.core.nglob`,
   and added this module to the reference documentation.
 
+### Fixed
+
+- Improve hash computation of a symbolic links in `stepup.core.hash`.
+
 
 ## [1.2.2] - 2024-05-16